+++ src/main/kotlin/com/jsrdev/med_api/domain/patient/PatientRepository.kt
@@ -1,4 +1,4 @@
-package com.jsrdev.med_api.patient
+package com.jsrdev.med_api.domain.patient
 
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
