+++ src/main/kotlin/com/jsrdev/med_api/domain/physician/PhysicianMapper.kt
@@ -1,7 +1,7 @@
-package com.jsrdev.med_api.physician
+package com.jsrdev.med_api.domain.physician
 
-import com.jsrdev.med_api.address.AddressMapper.toResponse
-import com.jsrdev.med_api.address.AddressMapper.updateFrom
+import com.jsrdev.med_api.domain.address.AddressMapper.toResponse
+import com.jsrdev.med_api.domain.address.AddressMapper.updateFrom
 
 object PhysicianMapper {
     fun Physician.toResponse() =
