+++ src/main/kotlin/com/jsrdev/med_api/domain/physician/PhysicianRequest.kt
@@ -1,13 +1,15 @@
-package com.jsrdev.med_api.patient
+package com.jsrdev.med_api.domain.physician
 
 import com.fasterxml.jackson.annotation.JsonAlias
-import com.jsrdev.med_api.address.AddressData
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize
+import com.jsrdev.med_api.domain.address.AddressData
 import jakarta.validation.Valid
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
+import jakarta.validation.constraints.Pattern
 
-data class PatientRequest(
+data class PhysicianRequest(
     @field:NotBlank(message = "Name must not be empty.")
     val name: String,
 
@@ -17,16 +19,21 @@ data class PatientRequest(
     @field:Email(message = "Email must be a valid format.")
     val email: String,
 
-    @JsonAlias("document_identity", "documentIdentity", "document")
-    @field:NotBlank(message = "Document identity must not be empty.")
-    val documentIdentity: String,
+    @field:NotBlank(message = "Document must not be empty.")
+    @field:Pattern(regexp = "\\d{4,9}", message = "Document must contain between 4 and 9 digits.")
+    val document: String,
 
-    @JsonAlias("phone_number", "phoneNumber")
+    @JsonAlias("phoneNumber", "phone_number")
     @field:NotBlank(message = "Phone number must not be empty.")
     val phoneNumber: String,
+
+    @JsonDeserialize(using = SpecialtyDeserializer::class)
+    @field:NotNull(message = "Specialty must not be null.")
+    val specialty: Specialty,
 
     @JsonAlias("address", "addressData", "address_data")
     @field:NotNull(message = "Address data must not be null.")
     @field:Valid
     val addressData: AddressData
 )
+
