+++ src/main/kotlin/com/jsrdev/med_api/domain/physician/UpdatePhysician.kt
@@ -1,13 +1,14 @@
-package com.jsrdev.med_api.patient
+package com.jsrdev.med_api.domain.physician
 
 import com.fasterxml.jackson.annotation.JsonAlias
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
-import com.jsrdev.med_api.address.AddressData
-import com.jsrdev.med_api.core.NonEmptyLongDeserializer
+import com.jsrdev.med_api.domain.address.AddressData
+import com.jsrdev.med_api.domain.core.NonEmptyLongDeserializer
 import jakarta.validation.Valid
 import jakarta.validation.constraints.NotNull
+import jakarta.validation.constraints.Pattern
 
-data class UpdatePatient(
+data class UpdatePhysician(
     @JsonDeserialize(using = NonEmptyLongDeserializer::class)
     @field:NotNull(message = "Id must not be null.")
     val id: Long,
@@ -16,8 +17,8 @@ data class UpdatePatient(
 
     val avatar: String?,
 
-    @JsonAlias("phone_number", "phoneNumber")
-    val phoneNumber: String?,
+    @field:Pattern(regexp = "\\d{4,9}", message = "Document must contain between 4 and 9 digits.")
+    val document: String?,
 
     @JsonAlias("address", "addressData", "address_data")
     @field:Valid
